% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Calculate odds ratios and confidence intervals}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{A vector of coefficient estimates}

\item{se}{A vector of standard errors}

\item{siglevel}{The significance level}

\item{roundto}{The number of decimal places}
}
\value{
Odds ratios and confidence intervals
}
\description{
Calculate odds ratios and confidence intervals
}
\examples{
OR_95CI(1.00551, 0.25341, 0.05, 3)
}
